Update to ability to apply auto dq and pen before results are then processed, successful implementation of bulsca scoring for each major event type

+++ app/DTO/RankedResult.php
@@ -23,8 +23,8 @@ class RankedResult extends Result
      */
     public function __construct(
         public int $id,
-        public int $resolvedResult,
-        public string|int|null $result,
+        public float $resolvedResult,
+        public string|float|null $result,
         public Entity $entity,
         public Event $event,
         public int $position,
